set up api routes
