+++ app.py
@@ -1,5 +1,6 @@
 from flask import *
 from api.attractions import attractions
+from api.user import user
 from flask_cors import CORS
 
 app=Flask(__name__)
@@ -9,6 +10,7 @@ app.config["JSON_AS_ASCII"]=False
 app.config["TEMPLATES_AUTO_RELOAD"]=True
 #register blueprint
 app.register_blueprint(attractions, url_prefix='/api')
+app.register_blueprint(user, url_prefix='/api')
 
 # Pages
 @app.route("/")
